faulthandler: avoid accessing sys.stderr.encoding
Fixes a pytest-xdist regression after
762bb61562d1300eedeb80be2ec2fb8150b3cc3f (not yet released).

pytest-xdist patches sys.stderr with an object which doesn't have
`encoding`. Strictly speaking, this should be fixed there (or more
precisely, in execnet), but it will drop support for older versions
which don't want.

But in any case, the fix turns out to simplify the code, using FD
support added in Python 3.5, so it's good anyway!

Refs: https://github.com/pytest-dev/pytest-xdist/pull/900